Minor changes xD
