+++ src/lib.rs
@@ -1,14 +1,24 @@
+pub mod api {
+    pub mod deployment;
+    pub mod device;
+    pub mod results;
+    pub mod run;
+}
 pub mod lib {
     pub mod wasmtime;
     pub mod wasmtime_imports;
     pub mod zeroconf;
-    pub mod api;
     pub mod constants;
     pub mod configuration;
     pub mod logging;
     pub mod deployment;
+    pub mod utils;
 }
 pub mod structs {
     pub mod device;
     pub mod request_entry;
+    pub mod deployment_orchestrator;
+    pub mod openapi;
+    pub mod module_orchestrator;
+    pub mod deployment_supervisor;
 }
