+++ src/main.rs
@@ -8,15 +8,19 @@
 //! - Registers the device with Zeroconf (mDNS/Bonjour)
 //! - Spawns a background worker thread for executing WebAssembly tasks asynchronously
 
+use actix_web::web;
 use actix_web::{App, HttpServer, web::Data};
 use actix_cors::Cors;
 use log::info;
 use parking_lot::Mutex;
+use supervisor::api::deployment::{deployment_create, deployment_delete, deployment_get};
+use supervisor::api::device::{register_orchestrator, thingi_description, thingi_health, wasmiot_device_description};
+use supervisor::api::results::{get_module_result, request_history_list, request_history_list_1};
+use supervisor::api::run::{run_module_function, run_module_function_3};
+use supervisor::structs::deployment_supervisor::Deployment;
 use std::sync::Arc;
-use supervisor::lib::{api, zeroconf, constants};
-use supervisor::lib::constants::DEPLOYMENTS_FOLDER;
-use supervisor::lib::deployment::Deployment;
-use supervisor::lib::api::DEPLOYMENTS;
+use supervisor::lib::{zeroconf, constants};
+use supervisor::lib::constants::{DEPLOYMENTS, DEPLOYMENTS_FOLDER};
 
 /// Main entry point for the supervisor service.
 ///
@@ -136,7 +140,95 @@ async fn main() -> std::io::Result<()> {
             actix_web::middleware::Logger::default()
         )
         .app_data(Data::new(zc_arc.clone()))  // Pass the Zeroconf instance to the app
-        .configure(api::configure_routes)
+        .service(
+                web::resource("/.well-known/wasmiot-device-description")
+                    .name("/.well-known/wasmiot-device-description")
+                    .route(web::get().to(wasmiot_device_description))
+            )
+
+            // Returns W3C WoT Thing Description for this device
+            .service(
+                web::resource("/.well-known/wot-thing-description")
+                    .name("/.well-known/wot-thing-description")
+                    .route(web::get().to(thingi_description))
+            )
+
+            // Health check for device (CPU, memory, network)
+            .service(
+                web::resource("/health")
+                    .name("/health")
+                    .route(web::get().to(thingi_health))
+            )
+            // Duplicate health route for compatibility
+            .service(
+                web::resource("//health")
+                    .name("//health")
+                    .route(web::get().to(thingi_health))
+            )
+
+            // Registers the active orchestrator URL to the device
+            .service(
+                web::resource("/register")
+                    .name("/register")
+                    .route(web::post().to(register_orchestrator))
+            )
+
+            // Fetch result files generated by module execution
+            .service(
+                web::resource("/module_results/{deployment_id}/{module_name}/{filename}")
+                    .name("/module_results/{deployment_id}/{module_name}/{filename}")
+                    .route(web::get().to(get_module_result))
+            )
+
+            // Fetch execution history (entire list or single entry by ID)
+            .service(
+                web::resource("/request-history/{request_id}")
+                    .name("/request-history/{request_id}")
+                    .route(web::get().to(request_history_list))
+            )
+            .service(
+                web::resource("/request-history")
+                    .name("/request-history")
+                    .route(web::get().to(request_history_list_1))
+            )
+
+            // Serve result file produced by specific function execution
+            .service(
+                web::resource("/{deployment_id}/modules/{module_name}/{function_name}/{filename}")
+                    .name("/{deployment_id}/modules/{module_name}/{function_name}/{filename}")
+                    .route(web::get().to(run_module_function))
+            )
+
+            // Run a module function (GET: immediate execution, no input files)
+            // Run a module function (POST: allows input files via multipart)
+            .service(
+                web::resource("/{deployment_id}/modules/{module_name}/{function_name}")
+                    .name("/{deployment_id}/modules/{module_name}/{function_name}")
+                    .route(web::get().to(run_module_function_3))
+                    .route(web::post().to(run_module_function_3))
+            )
+
+            // Delete an existing deployment by ID
+            .service(
+                web::resource("/deploy/{deployment_id}")
+                    .name("/deploy/{deployment_id}")
+                    .route(web::delete().to(deployment_delete))
+            )
+
+            // Get a list of all deployments currently active on this device
+            // Create a new deployment with modules and optional mount/config data
+            .service(
+                web::resource("/deploy")
+                    .name("/deploy")
+                    .route(web::get().to(deployment_get))
+                    .route(web::post().to(deployment_create))
+            )
+            // This is needed for current version of orchestrator for some reason
+            .service(
+                web::resource("//deploy")
+                    .name("//deploy")
+                    .route(web::post().to(deployment_create))
+            )
     })
     .bind(("0.0.0.0", port))?;
     info!("Starting supervisor service at http://{}:{}/", host, port);
